Fix bug of integer overflow
For the original code, there is an potential overflow when q < 32 but
(1 << q) * (r+3) > 2^32.
because (1 << q) < 2^32, compiler (like gcc) will use int to store it,
so does (r+3). so int*int will overflow when result > 2^32

To fix this, simply cast data type of (1 << q) or (r+3) to size_t.

+++ src/quotient-filter.c
@@ -387,7 +387,7 @@ void qf_clear(quotient_filter *qf)
 
 size_t qf_table_size(uint32_t q, uint32_t r)
 {
-    size_t bits = (1 << q) * (r + 3);
+    size_t bits = (size_t)(1 << q) * (r + 3);
     size_t bytes = bits / 8;
     return (bits % 8) ? (bytes + 1) : bytes;
 }
